% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epvclust.R
\name{epvdriver}
\alias{epvdriver}
\title{Main function of epvclust.
Compute hierchical clustering and calculate AU-values and BP-values for each cluster.}
\usage{
epvdriver(filename, nboot = 10, r = c(50, 75, 100, 125, 150),
  method.hclust = "average", transpose = 0)
}
\arguments{
\item{filename}{path he of the file cointaining the dataset (it must be in CSV format)}

\item{nboot}{number of bootstrap replicate to do for each bootstrap session}

\item{method.hclust}{clustering linkage method}

\item{transpose}{if 0 the objects are on columns}
}
\value{
hierachical clustering with AU-values and BP-values
}
\description{
Main function of epvclust.
Compute hierchical clustering and calculate AU-values and BP-values for each cluster.
}

